Remove FormatFilterPlanner
