+++ marklogic/src/main/scala/quasar/physical/marklogic/qscript/Planner.scala
@@ -85,7 +85,7 @@ sealed abstract class PlannerInstances0 extends PlannerInstances1 {
   ): Planner[F, FMT, Const[ShiftedRead[ADir], ?], J] =
     new ShiftedReadDirPlanner[F, FMT, J]
 
-  implicit def qScriptCore[F[_]: Monad: QNameGenerator: PrologW: MonadPlanErr: Xcc, FMT: SearchOptions: FormatFilterPlanner, T[_[_]]: BirecursiveT](
+  implicit def qScriptCore[F[_]: Monad: QNameGenerator: PrologW: MonadPlanErr: Xcc, FMT: SearchOptions: FilterPlanner[T, ?], T[_[_]]: BirecursiveT](
     implicit
     SP : StructuralPlanner[F, FMT],
     QTP: Lazy[Planner[F, FMT, QScriptTotal[T, ?], T[EJson]]]
